requirements.txt
flask>=3.0.0
flask-login>=0.6.3
flask-bcrypt>=1.0.1
mysql-connector-python>=8.4.0
python-dotenv>=1.0.0
